requirements.txt
Flask
flask-restful
gunicorn
requests
wiotp-sdk
